// frontend/src/generated/api/models/AlgorithmType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const AlgorithmType = {
    SeasonalNaive24: 'seasonal_naive_24',
    SeasonalNaive168: 'seasonal_naive_168',
    Ridge: 'ridge',
    RandomForest: 'random_forest',
    HistGradientBoosting: 'hist_gradient_boosting'
} as const;
export type AlgorithmType = typeof AlgorithmType[keyof typeof AlgorithmType];


export function instanceOfAlgorithmType(value: any): boolean {
    for (const key in AlgorithmType) {
        if (Object.prototype.hasOwnProperty.call(AlgorithmType, key)) {
            if (AlgorithmType[key as keyof typeof AlgorithmType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function AlgorithmTypeFromJSON(json: any): AlgorithmType {
    return AlgorithmTypeFromJSONTyped(json, false);
}

export function AlgorithmTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): AlgorithmType {
    return json as AlgorithmType;
}

export function AlgorithmTypeToJSON(value?: AlgorithmType | null): any {
    return value as any;
}

export function AlgorithmTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): AlgorithmType {
    return value as AlgorithmType;
}