// frontend/src/generated/api/models/ExperimentStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ExperimentStatus = {
    Queued: 'queued',
    Running: 'running',
    Cancelling: 'cancelling',
    Cancelled: 'cancelled',
    Completed: 'completed',
    Failed: 'failed'
} as const;
export type ExperimentStatus = typeof ExperimentStatus[keyof typeof ExperimentStatus];


export function instanceOfExperimentStatus(value: any): boolean {
    for (const key in ExperimentStatus) {
        if (Object.prototype.hasOwnProperty.call(ExperimentStatus, key)) {
            if (ExperimentStatus[key as keyof typeof ExperimentStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ExperimentStatusFromJSON(json: any): ExperimentStatus {
    return ExperimentStatusFromJSONTyped(json, false);
}

export function ExperimentStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExperimentStatus {
    return json as ExperimentStatus;
}

export function ExperimentStatusToJSON(value?: ExperimentStatus | null): any {
    return value as any;
}

export function ExperimentStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): ExperimentStatus {
    return value as ExperimentStatus;
}