// frontend/src/generated/api/models/WeatherMode.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const WeatherMode = {
    W0: 'W0',
    W1: 'W1'
} as const;
export type WeatherMode = typeof WeatherMode[keyof typeof WeatherMode];


export function instanceOfWeatherMode(value: any): boolean {
    for (const key in WeatherMode) {
        if (Object.prototype.hasOwnProperty.call(WeatherMode, key)) {
            if (WeatherMode[key as keyof typeof WeatherMode] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WeatherModeFromJSON(json: any): WeatherMode {
    return WeatherModeFromJSONTyped(json, false);
}

export function WeatherModeFromJSONTyped(json: any, ignoreDiscriminator: boolean): WeatherMode {
    return json as WeatherMode;
}

export function WeatherModeToJSON(value?: WeatherMode | null): any {
    return value as any;
}

export function WeatherModeToJSONTyped(value: any, ignoreDiscriminator: boolean): WeatherMode {
    return value as WeatherMode;
}