// frontend/src/generated/api/models/ExperimentResponse.ts
export interface ExperimentResponse {
    /**
     * 
     * @type {Array<AlgorithmType>}
     * @memberof ExperimentResponse
     */
    algorithms: Array<AlgorithmType>;
    /**
     * 
     * @type {Date}
     * @memberof ExperimentResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ExperimentResponse
     */
    datasetVersionId: string;
    /**
     * 
     * @type {string}
     * @memberof ExperimentResponse
     */
    failureCode: string | null;
    /**
     * 
     * @type {string}
     * @memberof ExperimentResponse
     */
    failureDetail: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ExperimentResponse
     */
    finishedAt: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ExperimentResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ExperimentResponse
     */
    jobId: string;
    /**
     * 
     * @type {string}
     * @memberof ExperimentResponse
     */
    name: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof ExperimentResponse
     */
    resultManifest: { [key: string]: any; } | null;
    /**
     * 
     * @type {SensitivityMode}
     * @memberof ExperimentResponse
     */
    sensitivityMode: SensitivityMode;
    /**
     * 
     * @type {Date}
     * @memberof ExperimentResponse
     */
    startedAt: Date | null;
    /**
     * 
     * @type {ExperimentStatus}
     * @memberof ExperimentResponse
     */
    status: ExperimentStatus;
    /**
     * 
     * @type {WeatherMode}
     * @memberof ExperimentResponse
     */
    weatherMode: WeatherMode;
}