// frontend/src/generated/api/models/JobResponse.ts
export interface JobResponse {
    /**
     * 
     * @type {number}
     * @memberof JobResponse
     */
    attempt: number;
    /**
     * 
     * @type {Array<JobAttemptResponse>}
     * @memberof JobResponse
     */
    attempts: Array<JobAttemptResponse>;
    /**
     * 
     * @type {Date}
     * @memberof JobResponse
     */
    cancelRequestedAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof JobResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof JobResponse
     */
    errorCode: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobResponse
     */
    errorDetail: string | null;
    /**
     * 
     * @type {Date}
     * @memberof JobResponse
     */
    finishedAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof JobResponse
     */
    heartbeatAt: Date | null;
    /**
     * 
     * @type {string}
     * @memberof JobResponse
     */
    id: string;
    /**
     * 
     * @type {JobType}
     * @memberof JobResponse
     */
    jobType: JobType;
    /**
     * 
     * @type {number}
     * @memberof JobResponse
     */
    maxAttempts: number;
    /**
     * 
     * @type {number}
     * @memberof JobResponse
     */
    progressPct: number;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof JobResponse
     */
    result: { [key: string]: any; } | null;
    /**
     * 
     * @type {Date}
     * @memberof JobResponse
     */
    startedAt: Date | null;
    /**
     * 
     * @type {JobStatus}
     * @memberof JobResponse
     */
    status: JobStatus;
    /**
     * 
     * @type {Date}
     * @memberof JobResponse
     */
    updatedAt: Date;
}